+++ lib/onboarding_flow.dart
@@ -1,5 +1,4 @@
 library onboarding_flow;
 
-export 'onboarding/button_params.dart';
 export 'onboarding/onboarding.dart';
 export 'onboarding/onboarding_page.dart';