+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Title, description and child added to page.
+* Removed button params.
+
 ## 0.0.2
 
 * Examples added